[apptools] Nominate CI tests for apptools
Impacted tests(approved): new 0, update 52, delete 0
Unit test platform: [Android]
Unit test result summary: pass 47, fail 5, block 0
