+++ apptools/apptools-android-tests/apptools/build.py
@@ -37,15 +37,6 @@ import shutil
 
 class TestCrosswalkApptoolsFunctions(unittest.TestCase):
 
-    def test_build_normal(self):
-        comm.setUp()
-        comm.create(self)
-        os.chdir('org.xwalk.test')
-        buildcmd = comm.HOST_PREFIX + comm.PackTools + "crosswalk-app build"
-        comm.build(self, buildcmd)
-        comm.run(self)
-        comm.clear("org.xwalk.test")
-
     def test_build_release(self):
         comm.setUp()
         comm.create(self)
