+++ apptools/apptools-android-tests/apptools/create_with_target_platform.py
@@ -35,19 +35,6 @@ import comm
 
 class TestCrosswalkApptoolsFunctions(unittest.TestCase):
 
-    def test_create_with_platform_android(self):
-        comm.setUp()
-        os.chdir(comm.XwalkPath)
-        comm.clear("org.xwalk.test")
-        cmd = comm.HOST_PREFIX + comm.PackTools + "crosswalk-app create org.xwalk.test" + comm.MODE + " --platform=android --android-crosswalk=" + comm.crosswalkzip
-        packstatus = os.system(cmd)
-        self.assertEquals(packstatus, 0)
-        os.chdir('org.xwalk.test')
-        buildcmd = comm.HOST_PREFIX + comm.PackTools + "crosswalk-app build"
-        comm.build(self, buildcmd)
-        comm.run(self)
-        comm.clear("org.xwalk.test")
-
     def test_create_with_platform_ios(self):
         comm.setUp()
         os.chdir(comm.XwalkPath)
