+++ apptools/apptools-ios-tests/apptools/manifest_display.py
@@ -37,21 +37,6 @@ import json
 
 class TestCrosswalkApptoolsFunctions(unittest.TestCase):
 
-    def test_display_fullscreen(self):
-        comm.setUp()
-        comm.create(self)
-        os.chdir('org.xwalk.test')
-        jsonfile = open(comm.ConstPath + "/../tools/org.xwalk.test/app/manifest.json", "r")
-        jsons = jsonfile.read()
-        jsonfile.close()
-        jsonDict = json.loads(jsons)
-        jsonDict["display"] = "fullscreen"
-        json.dump(jsonDict, open(comm.ConstPath + "/../tools/org.xwalk.test/app/manifest.json", "w"))
-        buildcmd = comm.PackTools + "crosswalk-app build"
-        buildstatus = os.system(buildcmd)
-        comm.clear("org.xwalk.test")
-        self.assertEquals(buildstatus, 0)
-
     def test_display_standalone(self):
         comm.setUp()
         comm.create(self)
