+++ apptools/apptools-ios-tests/tests.xml
@@ -8,7 +8,7 @@
           <test_script_entry>/opt/apptools-ios-tests/apptools/pkgName.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_build" purpose="iOS - Validate if project is built successfully with build command" subcase="2">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_build" purpose="iOS - Validate if project is built successfully with build command">
         <description>
           <test_script_entry>/opt/apptools-ios-tests/apptools/build.py</test_script_entry>
         </description>
@@ -18,17 +18,7 @@
           <test_script_entry>/opt/apptools-ios-tests/apptools/create.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_version" purpose="iOS - Validate if version command can display version information">
-        <description>
-          <test_script_entry>/opt/apptools-ios-tests/apptools/version.py</test_script_entry>
-        </description>
-      </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_target_platforms" purpose="iOS - Validate if platforms command can list available target platforms">
-        <description>
-          <test_script_entry>/opt/apptools-ios-tests/apptools/target_platforms.py</test_script_entry>
-        </description>
-      </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_create_with_target_platform" purpose="iOS - Validate project can be created with validate platform" subcase="4">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_create_with_target_platform" purpose="iOS - Validate project can be created with validate platform" subcase="3">
         <description>
           <test_script_entry>/opt/apptools-ios-tests/apptools/create_with_target_platform.py</test_script_entry>
         </description>
@@ -38,19 +28,14 @@
           <test_script_entry>/opt/apptools-ios-tests/apptools/build_path.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_package_id" purpose="iOS - Validate if package ID is used as expansion field" subcase="3">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_package_id" purpose="iOS - Validate if package ID is used as expansion field" subcase="2">
         <description>
           <test_script_entry>/opt/apptools-ios-tests/apptools/manifest_package_id.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_display" purpose="iOS - Validate if project can be built with valid display value" subcase="4">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_display" purpose="iOS - Validate if project can be built with valid display value" subcase="3">
         <description>
           <test_script_entry>/opt/apptools-ios-tests/apptools/manifest_display.py</test_script_entry>
-        </description>
-      </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_build_arguments" purpose="iOS - Validate if project can be built successfully with build arguments" subcase="3">
-        <description>
-          <test_script_entry>/opt/apptools-ios-tests/apptools/build_arguments.py</test_script_entry>
         </description>
       </testcase>
       <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_start_url" purpose="iOS - Validate if 'start_url' field can be set to 'http/https/blank' in the web manifest" subcase="3">
@@ -71,6 +56,16 @@
       <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_create_package" purpose="iOS - Validate if package can be created without a project">
         <description>
           <test_script_entry>/opt/apptools-ios-tests/apptools/create_package.py</test_script_entry>
+        </description>
+      </testcase>
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_app_basic" purpose="iOS - Validate if basic functions of crosswalk-app command can work fine" subcase="7">
+        <description>
+          <test_script_entry>/opt/apptools-ios-tests/apptools/CI/crosswalk_app_basic.py</test_script_entry>
+        </description>
+      </testcase>
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_basic" purpose="iOS - Validate if the fields in manifest can work fine" subcase="2">
+        <description>
+          <test_script_entry>/opt/apptools-ios-tests/apptools/CI/manifest_basic.py</test_script_entry>
         </description>
       </testcase>
     </set>
