+++ apptools/apptools-windows-tests/apptools/check_host.py
@@ -43,13 +43,5 @@ class TestCrosswalkApptoolsFunctions(unittest.TestCase):
         self.assertEquals(return_code, 0)
         self.assertNotIn("ERROR:", output[0].split('target windows')[1])
 
-    def test_check_host_windows(self):
-        comm.setUp()
-        os.chdir(comm.XwalkPath)
-        cmd = comm.HOST_PREFIX + comm.PackTools + "crosswalk-app check windows"
-        (return_code, output) = comm.getstatusoutput(cmd)
-        self.assertEquals(return_code, 0)
-        self.assertNotIn("ERROR:", output[0])
-
 if __name__ == '__main__':
     unittest.main()
