+++ apptools/apptools-windows-tests/apptools/init_manifest.py
@@ -49,21 +49,7 @@ class TestCrosswalkApptoolsFunctions(unittest.TestCase):
         with open(comm.ConstPath + "/../tools/org.xwalk.test/manifest.json") as json_file:
             data = json.load(json_file)
         comm.clear("org.xwalk.test")
-        self.assertEquals(data['xwalk_target_platforms'].strip(os.linesep), "android")
-
-    def test_init_manifest_windowsPlatforms(self):
-        comm.setUp()
-        os.chdir(comm.XwalkPath)
-        comm.clear("org.xwalk.test")
-        os.mkdir("org.xwalk.test")
-        cmd = comm.HOST_PREFIX + comm.PackTools + \
-            "crosswalk-app manifest " + \
-            comm.XwalkPath + "org.xwalk.test --platform=windows"
-        os.system(cmd)
-        with open(comm.ConstPath + "/../tools/org.xwalk.test/manifest.json") as json_file:
-            data = json.load(json_file)
-        comm.clear("org.xwalk.test")
-        self.assertEquals(data['xwalk_target_platforms'].strip(os.linesep), "windows")
+        self.assertEquals(data['xwalk_target_platforms'][0].strip(os.linesep), "android")
 
     def test_init_manifest_invalidPlatforms(self):
         comm.setUp()
@@ -76,27 +62,6 @@ class TestCrosswalkApptoolsFunctions(unittest.TestCase):
         return_code = os.system(cmd)
         comm.clear("org.xwalk.test")
         self.assertNotEquals(return_code, 0)
-
-    def test_init_manifest_packageid(self):
-        comm.setUp()
-        os.chdir(comm.XwalkPath)
-        comm.clear("org.xwalk.test")
-        os.mkdir("org.xwalk.test")
-        cmd = comm.HOST_PREFIX + comm.PackTools + \
-            "crosswalk-app manifest " + \
-            comm.XwalkPath + "org.xwalk.test --platform=windows --package-id=org.xwalk.test"
-        os.system(cmd)
-        with open(comm.ConstPath + "/../tools/org.xwalk.test/manifest.json") as json_file:
-            data = json.load(json_file)
-        updatecmd = comm.HOST_PREFIX + comm.PackTools + \
-            "crosswalk-app manifest " + \
-            comm.XwalkPath + "org.xwalk.test --platform=windows --package-id=org.test.foo"
-        os.system(updatecmd)
-        with open(comm.ConstPath + "/../tools/org.xwalk.test/manifest.json") as json_file_update:
-            updatedata = json.load(json_file_update)
-        comm.clear("org.xwalk.test")
-        self.assertEquals(data['xwalk_package_id'].strip(os.linesep), "org.xwalk.test")
-        self.assertEquals(updatedata['xwalk_package_id'].strip(os.linesep), "org.test.foo")
 
 if __name__ == '__main__':
     unittest.main()
