+++ apptools/apptools-windows-tests/tests.xml
@@ -13,32 +13,12 @@
           <test_script_entry>/opt/apptools-windows-tests/apptools/pkgName.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_build" purpose="Windows - Validate if project is built successfully with build command">
-        <description>
-          <test_script_entry>/opt/apptools-windows-tests/apptools/build.py</test_script_entry>
-        </description>
-      </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_app_version" purpose="Windows - Validate if crosswalk-app version command can display version information">
-        <description>
-          <test_script_entry>/opt/apptools-windows-tests/apptools/crosswalk_app_version.py</test_script_entry>
-        </description>
-      </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_pkg_version" purpose="Windows - Validate if crosswalk-pkg version command can display version information">
-        <description>
-          <test_script_entry>/opt/apptools-windows-tests/apptools/crosswalk_pkg_version.py</test_script_entry>
-        </description>
-      </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_target_platforms" purpose="Windows - Validate if platforms command can list available target platforms">
-        <description>
-          <test_script_entry>/opt/apptools-windows-tests/apptools/target_platforms.py</test_script_entry>
-        </description>
-      </testcase>
       <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_build_path" purpose="Windows - Validate if project can be built with webapp path">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/build_path.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_package_id" purpose="Windows - Validate if package ID is used as expansion field" subcase="3">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_package_id" purpose="Windows - Validate if package ID is used as expansion field" subcase="2">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/manifest_package_id.py</test_script_entry>
         </description>
@@ -53,7 +33,7 @@
           <test_script_entry>/opt/apptools-windows-tests/apptools/manifest_app_name.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_check_host" purpose="Windows - Validate if check command can test host setup" subcase="2">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_check_host" purpose="Windows - Validate if check command can test host setup">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/check_host.py</test_script_entry>
         </description>
@@ -63,17 +43,17 @@
           <test_script_entry>/opt/apptools-windows-tests/apptools/manifest_file_name.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_create_package" purpose="Windows - Validate if package can be created without a project" subcase="22">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_create_package" purpose="Windows - Validate if package can be created without a project" subcase="16">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/create_package.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_display" purpose="Windows - Validate if project can be built with valid display value" subcase="4">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_display" purpose="Windows - Validate if project can be built with valid display value" subcase="3">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/manifest_display.py</test_script_entry>
         </description>
       </testcase>
-      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_init_manifest" purpose="Windows - Validate if the web manifest can be initialised and updated" subcase="4">
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_init_manifest" purpose="Windows - Validate if the web manifest can be initialised and updated" subcase="2">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/init_manifest.py</test_script_entry>
         </description>
@@ -86,6 +66,21 @@
       <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_start_url" purpose="Windows - Validate if 'start_url' field can be set to 'http/https/blank' in the web manifest" subcase="3">
         <description>
           <test_script_entry>/opt/apptools-windows-tests/apptools/manifest_start_url.py</test_script_entry>
+        </description>
+      </testcase>
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_app_basic" purpose="Windows - Validate if basic functions of crosswalk-app command can work fine" subcase="6">
+        <description>
+          <test_script_entry>/opt/apptools-windows-tests/apptools/CI/crosswalk_app_basic.py</test_script_entry>
+        </description>
+      </testcase>
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_app_basic" purpose="Windows - Validate if basic functions of crosswalk-pkg command can work fine" subcase="7">
+        <description>
+          <test_script_entry>/opt/apptools-windows-tests/apptools/CI/crosswalk_pkg_basic.py</test_script_entry>
+        </description>
+      </testcase>
+      <testcase component="Crosswalk App Tools/CLI" execution_type="auto" id="Crosswalk_manifest_basic" purpose="Windows - Validate if the fields in manifest can work fine" subcase="2">
+        <description>
+          <test_script_entry>/opt/apptools-windows-tests/apptools/CI/manifest_basic.py</test_script_entry>
         </description>
       </testcase>
     </set>
